requirements.txt
gpiozero
requests
python-dotenv
flask
